Move TruffleLoader in to package-utils
