+++ packages/package-utils/index.js
@@ -3,3 +3,4 @@ exports.updateGasEstimate = require("./updateGasEstimate").default;
 exports.DiscordAdapter = require("./adapters/discord").default;
 exports.SlackAdapter = require("./adapters/slack").default;
 exports.ConsoleAdapter = require("./adapters/console").default;
+exports.TruffleLoader = require("./TruffleLoader").default;
