+++ packages/reputation-miner/patricia-test.js
@@ -3,7 +3,7 @@
 import path from "path";
 import { fromAscii } from "web3-utils";
 
-import TruffleLoader from "./TruffleLoader";
+import { TruffleLoader } from "../package-utils";
 import ReputationMiner from "./ReputationMiner";
 
 const EtherRouter = artifacts.require("EtherRouter");
