+++ test/contracts-network/colony-network-recovery.js
@@ -7,7 +7,7 @@ import chai from "chai";
 import bnChai from "bn-chai";
 import path from "path";
 
-import TruffleLoader from "../../packages/reputation-miner/TruffleLoader";
+import { TruffleLoader } from "../../packages/package-utils";
 import {
   forwardTime,
   makeReputationKey,
