+++ test/reputation-system/disputes-over-child-reputation.js
@@ -2,7 +2,7 @@ import path from "path";
 import chai from "chai";
 import bnChai from "bn-chai";
 
-import TruffleLoader from "../../packages/reputation-miner/TruffleLoader";
+import { TruffleLoader } from "../../packages/package-utils";
 import {
   forwardTime,
   checkErrorRevertEthers,
