+++ test/reputation-system/happy-paths.js
@@ -6,7 +6,7 @@ import { ethers } from "ethers";
 import chai from "chai";
 import bnChai from "bn-chai";
 
-import TruffleLoader from "../../packages/reputation-miner/TruffleLoader";
+import { TruffleLoader } from "../../packages/package-utils";
 import {
   forwardTime,
   submitAndForwardTimeToDispute,
