+++ test/reputation-system/reputation-mining-client/client-auto-functionality.js
@@ -4,7 +4,7 @@ import path from "path";
 import chai from "chai";
 import bnChai from "bn-chai";
 
-import TruffleLoader from "../../../packages/reputation-miner/TruffleLoader";
+import { TruffleLoader } from "../../../packages/package-utils";
 
 import { DEFAULT_STAKE, MINING_CYCLE_DURATION, CHALLENGE_RESPONSE_WINDOW_DURATION } from "../../../helpers/constants";
 import {
