+++ test/reputation-system/types-of-disagreement.js
@@ -4,7 +4,7 @@ import { toBN } from "web3-utils";
 import chai from "chai";
 import bnChai from "bn-chai";
 
-import TruffleLoader from "../../packages/reputation-miner/TruffleLoader";
+import { TruffleLoader } from "../../packages/package-utils";
 import {
   forwardTime,
   checkErrorRevert,
